Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 int main(){
-    int n=3.14;
-    scanf("%d",&n);
-    printf("%d",n);
+    float n=3.14;
+    scanf("%f",&n);
+    printf("%f",n);
     return 0;
 }
